// protocols/generated/typescript/common/v1/pagination_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts,import_extension=js"
// @generated from file proto/mindclade/common/v1/pagination.proto (package mindclade.common.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file proto/mindclade/common/v1/pagination.proto.
 */
export const file_proto_mindclade_common_v1_pagination: GenFile = /*@__PURE__*/
  fileDesc("Cipwcm90by9taW5kY2xhZGUvY29tbW9uL3YxL3BhZ2luYXRpb24ucHJvdG8SE21pbmRjbGFkZS5jb21tb24udjEiNAoLUGFnZVJlcXVlc3QSEgoKcGFnZV90b2tlbhgBIAEoCRIRCglwYWdlX3NpemUYAiABKA0iJwoMUGFnZVJlc3BvbnNlEhcKD25leHRfcGFnZV90b2tlbhgBIAEoCUJKWkhnaXRodWIuY29tL21pbmRjbGFkZS9taW5kY2xhZGUvcHJvdG9jb2xzL2dlbmVyYXRlZC9nby9jb21tb24vdjE7Y29tbW9udjFiBnByb3RvMw");

/**
 * Opaque pagination cursor and bounded page size convention.
 *
 * @generated from message mindclade.common.v1.PageRequest
 */
export type PageRequest = Message<"mindclade.common.v1.PageRequest"> & {
  /**
   * @generated from field: string page_token = 1;
   */
  pageToken: string;

  /**
   * @generated from field: uint32 page_size = 2;
   */
  pageSize: number;
};

/**
 * Describes the message mindclade.common.v1.PageRequest.
 * Use `create(PageRequestSchema)` to create a new message.
 */
export const PageRequestSchema: GenMessage<PageRequest> = /*@__PURE__*/
  messageDesc(file_proto_mindclade_common_v1_pagination, 0);

/**
 * @generated from message mindclade.common.v1.PageResponse
 */
export type PageResponse = Message<"mindclade.common.v1.PageResponse"> & {
  /**
   * @generated from field: string next_page_token = 1;
   */
  nextPageToken: string;
};

/**
 * Describes the message mindclade.common.v1.PageResponse.
 * Use `create(PageResponseSchema)` to create a new message.
 */
export const PageResponseSchema: GenMessage<PageResponse> = /*@__PURE__*/
  messageDesc(file_proto_mindclade_common_v1_pagination, 1);